pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-dns-ips"
version = "0.1.0"
description = "Resolve Google IP CIDR ranges from published DNS TXT records"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "GetGoogleIPs" }
]
dependencies = [
  "dnspython>=2.6.1",
  "flask>=3.0.0"
]
keywords = ["google", "dns", "ip", "cidr", "network"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

[tool.setuptools.packages.find]
where = ["src"]
